+++ changelog.md
@@ -1,10 +1,15 @@
 # Changelog
 
-### v1.1.0 - 25/05/2025
+**v1.1.1 - 25/05/2025**
+
+- Fixed colorblind mode on infoboxes not working correctly in Collection mode
+- Removed misleading help cursor in area above hotspots on Artefacts page
+
+**v1.1.0 - 25/05/2025**
 
 - Added settings menu and colorblind mode toggle
 - Improved menu behavior
 
-### v1.0.0 - 23/05/2025
+**v1.0.0 - 23/05/2025**
 
 - Released Arch Journal v1